# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Nemo Server API
  version: 1.0.0
  description: |
    Privacy-first cognitive AI platform providing real-time transcription,
    LLM chat, semantic search, and ML predictions.

    ## Authentication
    All protected endpoints require a Bearer token obtained from `/api/auth/login`.

    ## Rate Limiting
    Authentication endpoints are rate-limited to 10 requests per minute.
  contact:
    email: api@nemoserver.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.nemoserver.dev
    description: Production (if deployed)
...
security:
  - BearerAuth: []
paths:
...
  /api/auth/login:
    post:
      summary: Authenticate user
      description: Authenticate with username and password to receive session token.
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 50
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  session_token:
                    type: [string, "null"]
                  csrf_token:
                    type: [string, "null"]
                  username:
                    type: string
        '401':
          description: Invalid credentials
  /api/auth/logout:
    post:
      summary: End session
      description: Invalidate the current session token.
      tags: [Authentication]
      responses:
        '200':
          description: Logout successful
        '204':
          description: Session ended (no content)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Session token from /api/auth/login
    
    CSRFToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF token for mutating requests
...